+++ tests/test_sender_orm.py
@@ -19,24 +19,29 @@ class BasicTest(TestCase):
             password="testpassword"
         )
 
+        self.identifier1 = 'afdasfjasd3485'
+        self.identifier2 = 'vrefr9jqaw9efj'
+
         self.webook_target = WebhookTarget.objects.create(
             owner=self.user,
             event=WEBHOOK_EVENTS[0],
+            identifier=self.identifier1,
             target_url="http://httpbin.org/post"
         )
         self.fail_target = WebhookTarget.objects.create(
             owner=self.user,
             event=WEBHOOK_EVENTS[1],
+            identifier=self.identifier2,
             target_url="http://httpbin.org/status/400"
         )
 
     def test_webhook(self):
 
         @webhook(event=WEBHOOK_EVENTS[0])
-        def basic(owner):
+        def basic(owner, identifier):
             return {"what": "me worry?"}
 
-        results = basic(owner=self.user)
+        results = basic(owner=self.user, identifier=self.identifier1)
 
         self.assertEqual(results['what'], "me worry?")
 
@@ -49,10 +54,10 @@ class BasicTest(TestCase):
     def test_failed_webhook(self):
 
         @webhook(event=WEBHOOK_EVENTS[1])
-        def basic(owner):
+        def basic(owner, identifier):
             return {"what": "me worry?"}
 
-        results = basic(owner=self.user)
+        results = basic(owner=self.user, identifier=self.identifier2)
 
         self.assertEqual(results['what'], "me worry?")
 
@@ -63,7 +68,7 @@ class BasicTest(TestCase):
     def test_event_dkwarg(self):
 
         @webhook(number=123)
-        def basic(owner):
+        def basic(owner, identifier):
             return {"what": "me worry?"}
 
         self.assertRaises(TypeError, basic, self.user)
@@ -76,3 +81,15 @@ class BasicTest(TestCase):
 
         self.assertRaises(TypeError, basic)
 
+    def test_identifier_kwarg(self):
+
+        @webhook(event=WEBHOOK_EVENTS[0])
+        def basic(owner='pydanny'):
+            return {"what": "me worry?"}
+
+        with self.assertRaises(TypeError):
+            basic(owner='danny')
+
+    def test_cant_find_webhook_target(self):
+        # TODO
+        pass